requirements.txt
langchain-community>=0.0.10
langchain-core>=0.1.0
langchain-openai>=0.0.5
langgraph>=0.0.10
typing-extensions>=4.8.0
pydantic>=2.0.0
tavily-python>=0.1.8
python-dotenv>=1.0.0
chromadb>=0.4.0
tiktoken>=0.5.0
gradio